Keep opera services in an accessible location.
